% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMemo_script.R
\name{freqTable}
\alias{freqTable}
\title{Generate a frequency table}
\usage{
freqTable(
  inputds,
  variables,
  varnames = variables,
  n = TRUE,
  percent = TRUE,
  perc_digits = 1,
  bg = "aliceblue"
)
}
\arguments{
\item{inputds}{A dataset with one row per unit (i.e. a Step 2 dataset).}

\item{variables}{A vector of names indicating the variables you would like to
be cross-tabulated.}

\item{varnames}{A named \code{list} giving key-value pairs for how you would
like the \code{variables} to be labeled on the table.}

\item{n}{Logical. Should frequencies be displayed?}

\item{percent}{Logical. Should percents be displayed?}

\item{perc_digits}{A number indicating how many decimal places should be
displayed for the percentages.}

\item{bg}{The background color for the columns containing the calculated
statistics. Can be given as either a name (e.g. \code{"dodgerblue4"}) or
using the \code{\link[grDevices]{rgb}} function.}
}
\value{
A \code{\link[flextable]{flextable}} object.
}
\description{
The function \code{freqTable} is somewhat analagous to \code{PROC FREQ} in
SAS. The functionality is slightly different and more limited, but the
results are aesthetically similar.
}
\examples{
# Simple cross-tabulation using default options
freqTable(inputds    = mtcars,
          variables  = c('cyl', 'gear', 'carb'))

# Using options to customize table further
freqTable(inputds    = mtcars,
          variables  = c('cyl', 'gear', 'carb'),
          varnames   = list(cyl = 'Cylinders', gear = 'Forward gears', carb = 'Carburetors'),
          n          = FALSE,
          perc_digits = 0,
          bg         = rgb(.2, .1, .2, .3))
}
