% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMemo_script.R
\name{loadSASdf}
\alias{loadSASdf}
\title{Load SAS datasets}
\usage{
loadSASdf(
  rootdir,
  dirabb = "SD7",
  suffix = "",
  s1dat,
  s1names = NULL,
  s2dat,
  s2names = NULL
)
}
\arguments{
\item{rootdir}{The root directory for all programs and data. This directory
should contain subfolders for programs and SAS datasets (e.g. "prg" and
"sd7" folders).}

\item{dirabb}{The name of the directory (within the root directory) where the SAS datasets are stored.}

\item{suffix}{A suffix to add to the end of all the filenames (before the
file extension). For example, if the doc block lists the uncompressed
datasets and you would like to add "_uncomp" to all the filenames, setting
\code{suffix = "_uncomp"} would change "MyData.sas7bdat" to
"MyData_uncomp.sas7bdat".}

\item{s1dat}{A vector containing the names of all the Step 1 datasets. \emph{Must
end in .sas7bdat and be uncompressed.}}

\item{s1names}{A vector of names for the Step 1 datasets. \emph{Must be the
same length as \code{s1dat}}}

\item{s2dat}{The name of the Step 2 dataset. \emph{Must end in .sas7bdat and be uncompressed.}}

\item{s2names}{A vector of names for the Step 2 datasets. \emph{Must be the
same length as \code{s2dat}}}
}
\value{
A named \code{list} containing all the Step 1 and Step 2 datasets.
}
\description{
The \code{loadSASdf} function reads in the Step 1 and Step 2 files specified
in the doc block and automatically assigns them names (unless names are
user-specified).
}
