% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMemo_script.R
\name{makeQCLog}
\alias{makeQCLog}
\title{Generate the QC log}
\usage{
makeQCLog(
  rootdir,
  rawprg,
  rawdat,
  s1prg,
  s1dat,
  s2prg,
  s2dat,
  s2mac,
  prgmer_raw,
  prgmer_s1,
  prgmer_s2,
  reviewers,
  QC_date
)
}
\arguments{
\item{rootdir}{The root directory for all programs and data. This directory
should contain subfolders for programs and SAS datasets (e.g. "prg" and
"sd7" folders).}

\item{rawprg}{The name of the raw step program.}

\item{rawdat}{A vector containing the names of all the raw datasets.}

\item{s1prg}{The name of the Step 1 program.}

\item{s1dat}{A vector containing the names of all the Step 1 datasets. \emph{Must
end in .sas7bdat and be uncompressed.}}

\item{s2prg}{The name of the Step 2 program.}

\item{s2dat}{The name of the Step 2 dataset. \emph{Must end in .sas7bdat and be uncompressed.}}

\item{s2mac}{The name of the Step 2 macro Excel worksheet. \emph{Must end in .xlsm.}}

\item{prgmer_raw}{The raw step programmer.}

\item{prgmer_s1}{The Step 1 programmer.}

\item{prgmer_s2}{The Step 2 programmer.}

\item{reviewers}{The names of the QC memo reviewers}

\item{QC_date}{The date of the QC memo review.}
}
\value{
A \code{\link[flextable]{flextable}} object.
}
\description{
This function organizes the information from the doc block at the beginning
of the template into a table detailing all the programs, datasets, and other
files covered in the QC memo.
}
