% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMemo_script.R
\name{makeSSTable}
\alias{makeSSTable}
\title{Create sample size table}
\usage{
makeSSTable(
  inputds,
  cohort_var = "cohort",
  context_var = "context",
  racode_var = "racode"
)
}
\arguments{
\item{inputds}{A dataset with one row per unit (i.e. a Step 2 dataset).}

\item{cohort_var}{The name of the cohort variable on \code{inputds}.}

\item{context_var}{The name of the context variable on \code{inputds}.}

\item{racode_var}{The name of the research group variable on \code{inputds}.}
}
\value{
A \code{\link[flextable]{flextable}} object.
}
\description{
The function \code{makeSSTable} creates the standard sample size table
included in all QC memos (i.e., cross-tabulated by cohort, context, and
research group).
}
