% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMemo_script.R
\name{step2Contents}
\alias{step2Contents}
\title{Generate list of Step 2 outcomes}
\usage{
step2Contents(rootdir, dirabb = "PRG", s2mac)
}
\arguments{
\item{rootdir}{The root directory for all programs and data. This directory
should contain subfolders for programs and SAS datasets (e.g. "prg" and
"sd7" folders).}

\item{dirabb}{The name of the directory (within the root directory) where the
Step 2 Excel worksheet is stored.}

\item{s2mac}{The name of the Step 2 macro Excel worksheet. \emph{Must end in .xlsm.}}
}
\value{
A \code{\link[flextable]{flextable}} object.
}
\description{
The function \code{step2Contents} scrapes the Step 2 Excel worksheet to print
a table showing the stem and label for all Step 2 outcomes.
}
