% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCMemo_script.R
\name{summaryTable}
\alias{summaryTable}
\title{Generate a summary table}
\usage{
summaryTable(
  inputds,
  class,
  var,
  varnames = c(class, var),
  n = TRUE,
  nmiss = TRUE,
  percent = TRUE,
  mean = TRUE,
  min = TRUE,
  max = TRUE,
  perc_digits = 1,
  digits = 1,
  bg = "aliceblue",
  date = FALSE
)
}
\arguments{
\item{inputds}{A dataset to be summarized.}

\item{class}{A vector of names indicating variables to be used for
classification.}

\item{var}{The name of the analysis variable.}

\item{varnames}{A named \code{list} giving key-value pairs for how you would
like the variables to be labeled on the table.}

\item{n}{Logical. Should frequencies be displayed?}

\item{nmiss}{Logical. Should number of missing values be displayed?}

\item{percent}{Logical. Should percents be displayed?}

\item{mean}{Logical. Should means be displayed?}

\item{min}{Logical. Should minima be displayed?}

\item{max}{Logical. Should maxima be displayed?}

\item{perc_digits}{A number indicating how many decimal places should be
displayed for the percentages.}

\item{digits}{A number indicating how many decimal places should be displayed
for the mean, minimum, and maximum}

\item{bg}{The background color for the columns containing the calculated
statistics. Can be given as either a name (e.g. \code{"dodgerblue4"}) or
using the \code{\link[grDevices]{rgb}} function.}

\item{date}{Logical. Is the analysis variable a date?}
}
\value{
A \code{\link[flextable]{flextable}} object.
}
\description{
The function \code{summaryTable} is roughly analogous to \code{PROC MEANS} in
SAS. The functionality is slightly more limited than \code{PROC MEANS}, but
the results are aesthetically similar.
}
\examples{
# Simple table with one class variable
summaryTable(inputds     = mtcars,
             class       = 'cyl',
             var         = 'mpg',
             varnames    = list(cyl = 'Cylinders', mpg = 'Miles per gallon'))

# Two class variables and customization through additional parameters
summaryTable(inputds     = mtcars,
             class       = c('cyl', 'gear'),
             var         = 'mpg',
             varnames    = list(cyl  = 'Cylinders',
                                gear = 'Number of forward gears',
                                mpg  = 'Miles per gallon'),
             nmiss       = FALSE,
             perc_digits = 0,
             digits      = 2,
             bg          = 'olivedrab1')
}
